各酸素分圧ごとでデータを入力できるように

+++ src/rheed_segmentation/train.py
@@ -28,7 +28,7 @@ class LossComputer:
         self, preds: Tensor, targets: Tensor | dict[str, Tensor]
     ) -> tuple[Tensor, dict[str, float]]:
         if isinstance(targets, Tensor):
-            loss: Tensor = self.criterion(preds, targets)
+            loss: Tensor = self.criterion(preds, targets.long())
             return loss, {"all": loss.item()}
 
         if isinstance(targets, dict):
